chore: created ImageSelector component
